+++ examples/PySDM_examples/Arabas_et_al_2015/example_benchmark.py
@@ -2,7 +2,7 @@ import importlib
 
 from matplotlib import pyplot as plt
 from PySDM_examples.Arabas_et_al_2015 import Settings
-from PySDM_examples.utils.simulation_2D_prescribed_flow import Simulation, Storage
+from PySDM_examples.utils.kinematic_2d import Simulation, Storage
 
 import PySDM.backends.impl_numba.conf
 from PySDM import Formulae
