+++ examples/PySDM_examples/utils/kinematic_2d/gui_settings.py
