+++ examples/PySDM_examples/utils/kinematic_2d/make_default_product_collection.py
