+++ tests/smoke_tests/kinematic_2d/arabas_et_al_2015/test_gui_settings.py
@@ -1,6 +1,6 @@
 # pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring
 from PySDM_examples.Arabas_et_al_2015 import Settings
-from PySDM_examples.utils.simulation_2D_prescribed_flow.gui_settings import GUISettings
+from PySDM_examples.utils.kinematic_2d.gui_settings import GUISettings
 
 
 class TestGUISettings:
